wip: source

+++ scripts/dev/cli.py
@@ -7,14 +7,15 @@ from common import BUILD_DIR, THIRD_PARTY_DIR
 from doctor import doctor
 
 
-def clean():
-    if os.path.exists(THIRD_PARTY_DIR):
-        print(f"Deleting {THIRD_PARTY_DIR}")
-        shutil.rmtree(THIRD_PARTY_DIR)
-
+def clean(all: bool):
     if os.path.exists(BUILD_DIR):
         print(f"Deleting {BUILD_DIR}")
         shutil.rmtree(BUILD_DIR)
+
+    if all:
+        if os.path.exists(THIRD_PARTY_DIR):
+            print(f"Deleting {THIRD_PARTY_DIR}")
+            shutil.rmtree(THIRD_PARTY_DIR)
 
 
 def build():
@@ -35,8 +36,13 @@ def build():
 def cli():
     parser = argparse.ArgumentParser(description="Kero Development Tool.")
     commands = parser.add_subparsers(dest="command")
-    _clean_parser = commands.add_parser(
+    clean_parser = commands.add_parser(
         "clean", help="Clean the Kero development environment."
+    )
+    clean_parser.add_argument(
+        "--all",
+        action="store_true",
+        help="Also clean the third_party directory.",
     )
     _doctor_parser = commands.add_parser(
         "doctor", help="Check the Kero development environment."
@@ -44,7 +50,7 @@ def cli():
     _build_parser = commands.add_parser("build", help="Build the Kero.")
     args = parser.parse_args()
     if args.command == "clean":
-        clean()
+        clean(args.all)
     elif args.command == "doctor":
         doctor()
     elif args.command == "build":
